requirements.txt
uvicorn==0.29.0
fastapi==0.111.0
botbuilder-integration-aiohttp>=4.14.5
azure-common==1.1.28
azure-core==1.30.0
azure-cosmos==4.5.1
azure-identity==1.15.0
azure-keyvault-secrets==4.8.0
azure-search-documents==11.4.0
azure-storage-blob==12.19.0
azure-storage-file-datalake==12.14.0
blinker==1.7.0
cachetools==5.3.3
chardet==5.2.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
databricks==0.2
databricks-sql-connector==2.9.3
distro==1.9.0
dnspython==2.6.1
et-xmlfile==1.1.0
Flask==3.0.2
Flask-Cors==4.0.0
fonttools==4.49.0
google-api-core==2.18.0
google-auth==2.29.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonify==0.5
jsonpickle==1.4.2
jsonschema==4.21.1
jsonschema-path==0.3.2
jsonschema-spec==0.2.4
jsonschema-specifications==2023.7.1
kaleido==0.1.0
kiwisolver==1.4.5
lazy-object-proxy==1.10.0
lz4==4.3.3
Mako==1.3.2
MarkupSafe==2.1.5
matplotlib==3.8.3
more-itertools==10.2.0
motor==3.3.2
numpy==1.26.4
oauthlib==3.2.2
openai==1.12.0
openapi-core==0.18.2
openapi-schema-validator==0.6.2
openapi-spec-validator==0.7.1
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.13
openpyxl==3.1.2
packaging==23.2
pandas==2.1.4
parse==1.20.1
pathable==0.4.3
pillow==10.2.0
plotly==5.19.0
portalocker==2.8.2
prance==23.6.21.0
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
pyarrow==14.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.6.2
pydantic_core==2.16.3
pymongo==4.6.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.30.2
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rpds-py==0.18.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
semantic-kernel==0.5.1.dev0
sniffio==1.3.1
SQLAlchemy==1.4.52
tenacity==8.2.3
thrift==0.16.0
tiktoken==0.6.0
tqdm==4.66.2
tzdata==2024.1
Werkzeug==3.0.1
tabulate==0.9.0
python-dotenv==1.0.0
nltk
scikit-learn
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.1